requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-bigquery-storage
pandas
pyarrow
db-dtypes
matplotlib
seaborn
jupyter
apache-airflow
apache-airflow-providers-google
pytest
dvc
great-expectations
dvc-gs